# Conflicts:
#	src/main/java/web/mvc/dto/UserBuyDTO.java
#	src/main/java/web/mvc/repository/BuyMyPageRepository.java
#	src/test/java/web/mvc/StockTest.java

+++ src/main/java/web/mvc/controller/StockController.java
@@ -26,7 +26,7 @@ public class StockController {
 
     // 상품 등록 -> 상품 이미지 어떻게 하는지
     @PostMapping("/stock")
-    public ResponseEntity<?> insert(Long productSeq, Long stockGradeSeq, Long stockOrganicSeq, Long farmerSeq, @RequestBody StockDTO stockDTO) {
+    public ResponseEntity<?> insert(long productSeq, long stockGradeSeq, long stockOrganicSeq, long farmerSeq, @RequestBody StockDTO stockDTO) {
         log.info("Controller Product : {}", stockDTO);
 
         Stock stock = modelMapper.map(stockDTO, Stock.class);
@@ -64,7 +64,7 @@ public class StockController {
         System.out.println(stockDTO.getStockGradeSeq());
         System.out.println(stockDTO.getStockOrganicSeq());
 
-        int id = stockDTO.getStockSeq();
+        long id = stockDTO.getStockSeq();
 
         Stock stock = modelMapper.map(stockDTO, Stock.class);
