+++ src/main/java/web/mvc/repository/AuctionRepository.java
@@ -3,5 +3,5 @@ package web.mvc.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import web.mvc.domain.Auction;
 
-public interface AuctionRepository extends JpaRepository<Auction, Integer> {
+public interface AuctionRepository extends JpaRepository<Auction, Long> {
 }
